requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
certifi==2024.12.14
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.1
psycopg==3.2.3
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
yarl==1.18.3
